requirements.txt
numpy==1.24.3
pandas==1.5.3
scipy==1.10.1
scikit-learn==1.2.2
tqdm==4.62.3
pytorch==1.12.0 
torchvision==0.13.0
torchaudio==0.12.0
tqdm==4.65.0
wandb==0.16.6
yaml==0.2.5